working twin buffering
이중버퍼링 구현중

+++ HiwoongEngine/HiwoongEngine/Math/Vector2.cpp
@@ -14,6 +14,7 @@ namespace Hiwoong
 	Vector2::~Vector2()
 	{
 	}
+
 	Vector2 Vector2::operator+(const Vector2& other) const
 	{
 		return Vector2(x + other.x, y + other.y);
@@ -26,7 +27,7 @@ namespace Hiwoong
 	{
 		return Vector2(x * other.x, y * other.y);
 	}
-	Vector2 Vector2::operator*(float value) const
+	Vector2 Vector2::operator*(int value) const
 	{
 		return Vector2(x * value, y * value);
 	}
